+++ compose/src/main/kotlin/com/yourssu/handy/compose/textfield/TextField.kt
@@ -1,24 +1,15 @@
 package com.yourssu.handy.compose.textfield
 
-import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Spacer
-import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.height
-import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.wrapContentSize
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.remember
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.vector.ImageVector
-import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import com.yourssu.handy.compose.HandyTheme
 import com.yourssu.handy.compose.Text
 import com.yourssu.handy.compose.foundation.HandyTypography
-import com.yourssu.handy.compose.icons.HandyIcons
-import com.yourssu.handy.compose.icons.filled.Cancel
-
 @Composable
 fun TextField(
     value: String,
